[Feat] ObjectPool, Obserable, Singleton 관련 클래스 추가
