--- Assets/FolderColor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b725c631b97184544830dfadd6476140
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
