--- Assets/Patterns.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a0c4def6ae746a065827dbb4fde3e849
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
