+++ Assets/Scripts/Patterns/IPooedObject.cs.meta
