+++ Assets/Scripts/Patterns/Observable.cs.meta
