+++ Assets/Scripts/Patterns/Singleton.cs.meta
