--- Assets/Scripts/empty.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6f1450bf3caea7ccc8be12f28464395b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
